+++ src/extensions/bigbang/manifests.go
@@ -13,7 +13,7 @@ import (
 	"github.com/Masterminds/semver/v3"
 	fluxHelmCtrl "github.com/fluxcd/helm-controller/api/v2beta1"
 	fluxSrcCtrl "github.com/fluxcd/source-controller/api/v1"
-	"github.com/zarf-dev/zarf/src/types/extensions"
+	"github.com/zarf-dev/zarf/src/api/v1alpha1/extensions"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
