+++ zarf.schema.json
@@ -1,6 +1,6 @@
 {
   "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "$id": "https://github.com/zarf-dev/zarf/src/types/zarf-package",
+  "$id": "https://github.com/zarf-dev/zarf/src/api/v1alpha1/zarf-package",
   "$defs": {
     "BigBang": {
       "properties": {
@@ -1103,6 +1103,13 @@
     }
   },
   "properties": {
+    "apiVersion": {
+      "type": "string",
+      "enum": [
+        "zarf.dev/v1alpha1"
+      ],
+      "description": "The API version of the Zarf package."
+    },
     "kind": {
       "type": "string",
       "enum": [
@@ -1146,6 +1153,7 @@
   "additionalProperties": false,
   "type": "object",
   "required": [
+    "apiVersion",
     "kind",
     "components"
   ],
